[EMAIL] Changes in email [version] Upgrade to 1.2.6
