+++ src/framework/mailer/email-templates/welcome.ts
@@ -4,12 +4,11 @@ export default `
   </p>
   
   <p>
-  Thanks for joining {{siteName}}, This email is to inform you that you have joined {{siteName}} on {{ date }}, <a
-    href="{{activationUrl}}?token={{activationToken}}">Click Here</a> to verify your account.
+    Thanks for joining {{siteName}}, This email is to inform you that you have joined {{siteName}} on {{ date }}, <a target="_blank" href="{{activationUrl}}?token={{activationToken}}">Click Here</a> to verify your account.
   </p>
 
   <p>
-  Thanks <br />
-  {{ siteName }} Team.
+    Thanks <br />
+    {{ siteName }} Team.
   </p>
 `;
