+++ src/framework/mailer/index.ts
@@ -1,6 +1,7 @@
 let nodemailer = require('nodemailer');
+let handlebars = require("handlebars");
 
-export default async function name() {
+const defaultMailerInstance = async function ()  {
   let testAccount = await nodemailer.createTestAccount();
   let transporter = nodemailer.createTransport({
     host: "smtp.ethereal.email",
@@ -12,4 +13,27 @@ export default async function name() {
     }
   });
   return transporter;
+}
+
+export default async function mailerInstance() {
+  return await defaultMailerInstance();
+}
+
+export const sendEmail = async function (template: string, variables: any, credentials: any) {
+  let transporter: any = await mailerInstance();
+	let compiled = handlebars.compile(template);
+	let resultTemplate = compiled(variables);
+	try {
+		let send = await transporter.sendMail({
+			from : credentials.from,
+			to: credentials.to,
+			html: resultTemplate,
+			subject: credentials.subject
+		})
+		console.log("Message sent: %s", send.messageId);
+		console.log("Preview URL: %s", nodemailer.getTestMessageUrl(send));
+		return send;
+	} catch (e) {
+		console.log(`Failed sending email: ${e.message}`);
+	}
 }
